inheritance_updates
